test(async): isolate default header coverage
